+++ Private/NUnit/Publish-ResultsAndLogs.ps1
@@ -8,7 +8,7 @@ function Publish-ResultsAndLogs {
     )
 
     if($ImportResultsToCIServer) {
-      TeamCity-ImportNUnitReport "$AssemblyPath.$TestResultFilenamePattern.xml"
+      CI-ImportNUnitReport "$AssemblyPath.$TestResultFilenamePattern.xml"
     }
 
     # Tell teamcity to keep our test output logs as well. This could come in handy